Avoid coreference between current state and _last_published_data
Previously, all data source drivers replace the entire state on each
poll or push update (doctor). The DSE publish code relied on this
situation to optimize performance by skipping a copy when setting
_last_published_data. The coreference introduced did not matter
because the original state was never edited, only replaced.

But beginning with the webhook drivers
(monasca webhook and vitrage), the drivers edit the current state.
This a bug was introduced where sometimes an update would not be
published to messaging bus because of an incorrect _last_published_data
due to coreference.

This patch fixes the bug. In addition, the _last_published_data
elements are stored as set rather than list to save the need to
convert to set each time a diff is needed. The times when
conversion back to list is needed to publish on the messaging bus
(used for snapshot publish only)
is expected to be far fewer than the times the conversion to set is
needed.

A few additionalogging lines are also introbuced for improved
debuggability.

Change-Id: I332a844beaf8dec4450bb1ade6ce7e41cc3d4d91