Enable monasca in py2 tempest CI job
On bionic, there seems to be a kafka bug which causes some failures
on certain versions of JRE [1].
As a result, the JRE install invoked by Congress devstack plugin
causes Monasca devstack plugin to fail. Here we remove the JRE install
from Congress devstack plugin and enable monasca devstack plugin in
the CI job.

Note: Java is used to generate code for Congress, but not currently
required in Congress devstack install.

[1] https://stackoverflow.com/questions/36970622/kafka-unrecognized-vm-option-printgcdatestamps

Change-Id: Ibce028cb5635fe52275968beb5b4305f855ea851