move config init from congress_servers.py to bin/congress-server
The need to execute config init before importing harness in
congress_servers.py causes a style violation in hacking 1.1.0.

Here we move the initialization in bin/congress-server because it's
the first place that sees the command line options.

It also allows the work on upping hacking to 1.1.0 to go ahead.

Change-Id: I7fbc12195f5d79d7df8da3bb835f3f89c1244efc