Generic webhook processer for JsonIngester
Each table accepting webhook is configured with a jsonpath expression
to grab the unique ID from the webhook payload.

Each time a webhook comes in, the new data is written to the Postgres
database in jsonb format, replacing any previous webhook data with
the same ID.

The _key column type is changed from bigint to text in order to
accommodate custom keys in generic webhook processing.

PollingJsonIngester is renamed to JsonIngester because it is no
longer specialized to polling.

config format is changed slightly to clearly define which tables
are configured for polling and/or webhook.

A table that is updated through both poll and webhook is not
supported by this patch, but may be supported in the future.

Change-Id: I23c3d5c2c63fd8471868c2b37fd154eb2d1f9783
partially-implements: bp json-data-model