Run only smoke tests in ipv6 job
'congress-tempest-ipv6-only' job is to verify
that congress is able to listen and run on IPv6
env. We do not need to run all the tests in this
job. Running only smoke tests should be fine.

This commit define the 'congress-ipv6' tox env which
include smoke test only as of now but it can be extended
to include more IPv6 tests if any.

Change-Id: Idbd4c6ea362487dd93330adddb26abc89387a726
Story: #2005477
Task: #35885